Updates dependencies and Node version
Updates axios dependency to the latest version.

Updates Node.js version to v20 in ARM templates to ensure compatibility and take advantage of the latest features and security updates.

Sets the `FUNCTIONS_LOGGING_CONSOLE_LEVEL` environment variable to "Information" in ARM templates to control the verbosity of logging.

Uses `zlib` directly instead of `node:zlib`.

Converts the output of `zlib.gzipSync` to a `Buffer` before converting to a base64 string in order to ensure consistent encoding.

Removes unnecessary check of the `.log.gz` extension, and changes the path to include the extension directly.